Fix error when component has object as props
When attempting to render a component's props in the story source pane – when a component has an object for a prop like `style`, `previewProp()` was attempting to render a literal object with the key `content`, which causes a React error. Seems the intention of the original code was to render the literal characters `"{"` and `"}"` around `content`.