Mention TSX index file requirement in README
I was stuck getting Storybook working with TypeScript following the steps here. My components weren't showing up in the Storybook index. Diffing my project with the [demo repo](https://github.com/johot/storybook4-cra2-typescript-react-docgen-typescript-demo) I realized that my index file had the wrong name. Not sure where this requirement comes from, but it doesn't seem to work with `stories/index.tsx`, which is the likeliest name the user is going to give this file if the default is `stories/index.js`.