ci: remove `node_modules` cache
After walking through ~100 CircleCI runs I didn't find a single one during which `node_modules` cache was really used and resulted in reducing the `yarn install` phase to almost 0s.
However, saving and restoring this cache takes time on each run... so let's try to remove it and see what will happen.

It's possible that this cache is not working well as we are in a monorepo and we are caching only root `node_modules` so yarn still have to fetch some deps for subpackages