Use a custom `decorateStory` function in React
To fix issue with stories remounting when re-rendering. The issue was the way the `decorateStory` function created a new "bound" story function each time you called it.

This meant React wasn't able to recognize the fact that we were just changing props to the story rather than rendering a whole new component.

The fix is to do the "binding" via React context rather than by creating a new bound function on each call.

Also, tweaked the loaders implementation so React could avoid rebinding the decorated story function with the loaded value each time.