chore(addon): more `react` and `react-dom` to regular deps
Have both these packages as peerDep is causing a lot of `unmet peer dependency` warnings.
However, there is no need to worry about them because there is a webpack alias in `@storybook/core` so all import of react in addons are aliased to the one from `@storybook/core`.

For details, see https://github.com/storybookjs/storybook/issues/13194

I didn't use `*` as version range but the one used in `app/*` to have only 1 version in the monorepo `node_modules`.