Add `storyFilterFunction` to storyshots options
Exposing a filter function is more flexible for users and they can
finetune the matching to their liking. Initially I thought it could
deprecate the `storyNameRegex` and `storyKindRegex`, because we could
just create a default filter function which applies those two.

Our use case for this filter function is splitting the story shots based
on an arbitrary count of machines doing the tests in CI. But it will
also make it possible for people just to implement their test matching
to their liking.