Add documentation on usage with Vue CLI
Ref: https://cli.vuejs.org/guide/webpack.html#inspecting-the-project-s-webpack-config

If project is generated with Vue CLI, `vue-svg-loader` will require the use of the `chainWebpack` function which means `vue.config.js` wont be a fully resolved webpack config.  In that case the resolved webpack config can be found at `https://cli.vuejs.org/guide/webpack.html#inspecting-the-project-s-webpack-config` according to the above linked documentation.