When we do server-checks we naively assume that when requesting the resource, we'll get a 200 response only when the resource was accessible.
But in actuallity it's normal for a 200 statusCode to be returned but it's actually an authentication page.

I added an additional check to re-request the resource with an accept-header, and check the absence of a loginUrl field on the JSON response.

If the response is not valid JSON this additional check is ignored, and the previously behavior is preserved.