fix(angular): EXPERIMENTAL use module commonjs
This is a workaround to at least get the Angular example
app to compile. The startup of the example app fails
because it tries to load builders/start-storybook.js
that includes `import` statements which seems to be
incompatible with mjs if you don't explicitly set
`type: "module""` in the package.json of app/angular.

Setting `type: "module"` however disallows directory
path imports (e.g. rxjs/operators is incorrect and
it expects a js file in the import statement)