fix(angular): EXPERIMENTAL use webpack 5.70.0
Unfortunately, all the Angular packages use
a fix version for webpack (right now 5.70,0).
Webpack 5 is highly allergic to version mixing
and throws an error that does not seem to be
related to a dependency issue.

In this commit I set all versions in this repo
to 5.70.0 just to demonstrate that this is working.
I could not figure out another way to solve this.
I also tried the yarn "nohoist" feature but could
not get it working at all.

This is most likely only an issue in this repo and
user-land should not be affected by this conflict
as long as we offer a range that suits the
Angular fix version of webpack.