fix the build system.
Emotion changed to use exports map, and esbuild wasn't pivking the right entry anymore,
resulting in react not being detected, and thus not bundled, and thus missing at runtime!

I changed the tsup/esbuild config to prefer the ESM version, and this seems to have fixed the issue.

Many thanks to @Anderist for the assist!