Revert changes to `exports` fields of addons/renderers
All of Node, Webpack and Vite can `import('<pkg>/preview')` etc, as long as the `exports` map is there and setup like so. If the `exports` map is not there, then they can import `node_modules/<pkg>/preview.js` instead.

The issue occurred when an export map existed but Vite tried to import a path that wasn't defined in there.