Add proper Next.js framework support
NextJS projects should work out-of-the-box when Storybook is initialized.

1. Scenario: .babelrc or project babel.config.(m)js(x) file doesn't exist

In this case Next.js >= 12 uses swc to transpile the code. In Storybook
we will still use babel to transpile the code. We use the preset `next/babel`
and configure it to make the Next.js project work with Storybook. Any SWC/Next compiler
related options are ignored

2. Scenario: .babelrc or project babel.config.(m)js(x) file exist

Storybook will use the project's babel configuration for transpiling. All options, which
are passed to the `next/babel` prest are respected.