Add a test case a fix for #20958
A MDX file can reference a component or other non-story file with the same prefix as an MDX file, which leads to a matching entry in the cache but that entry will not have been processed yet. That's OK, we just assume that import is not a story file (rather than throwing).