Exclude decorators by default.
Due to reordering in https://github.com/storybookjs/storybook/pull/21182 the source decorator comes *before* any user decorator in `preview.js`. This means that we commonly trigger the bug in #21649.

Really it makes more sense to default this to true and people can opt out in cases where they want to show the decorator.