Angular: Implement singletons API to provide application-wide providers
In Storybook 7.0 we use the new bootstrapApplication API to bootstrap a middleman Storybook Standalone Module to the DOM. The module is configured in a way to respect your configured imports, providers, and declarations. This means also, that there is no root ngModule anymore. Previously you were able to add ModuleWithProviders, likely the result of a 'Module.forRoot()'-style call, to your 'imports' array of the moduleMetadata definition. This is now discouraged. Instead, you should use the new 'singletons' array of the moduleMetadata definition to add your application-wide providers.