Hopefully stabilize interactions e2e test
So… I discovered after a bunch of :computer-rage: ..
That the e2e test for interactions sometimes toggles the toolbar off.
When this happens it can’t click on the button to remount the story.
The toolbar disappears when playwright clicks on the addon tab, which appear like it refreshes storybook somehow, and then the toolbar disappears.
I cannot get playwright to do this locally, but I can see that something seems to be pressing t? Which toggles the toolbar.
What causes the refresh? what causes the toolbar to be toggle off? who knows?
I was trying to detect the presence of the toolbar to toggle back on, if needed,, and this is when things go weird.
Playwright tried to click the button, and can’t.. because it’s not visible.
So obviously I add a if statement, that if the button isn’t visible, it should do a procedure to ensure it will be, then try again.
Playwright tells me the button IS visible, but then when it tries to click says it isn’t. In addition using page.keyboard.press('t') doesn’t work to toggle the toolbar, very strange.
I even tried isWithinViewport()this suggests it’s in the viewport as well.