Update about page
The styling of 'Code' dramatically changed. Instead of using a fixed shade for the background, it now changes based on the theme. This improves visibility for users regardless of their chosen theme.

Lastly, the border color was adjusted to use the app's border color, creating more consistency in the application's look and feel.