Update storybook-test-runner-local-build-workflow.yml.mdx
When following this guide, there was an issue with the usage of the "wait-on" package when passing the parameter "tcp:6006" on the Github Actions VM. When the "wait-on" package doesn't explicitly specify the IP, it changes to the format "tcp:localhost:6006," which can cause problems if the localhost domain is not bound to the 127.0.0.1 IP in the CI VM environment.

To address this, I propose the following modification to the documentation:

Instead of using "tcp:6006" as the parameter, explicitly use "tcp:localhost:6006" instead. This will increase the likelihood of it functioning correctly in the CI VM environment.

Original:

```
wait-on tcp:6006
```

Proposed modification:

```
wait-on tcp:127.0.0.1:6006
```

By incorporating this change into the guide, the issue should be resolved, and it should work properly even in the CI environment.