Skip empty Vue 3 slot generation
We have this kind of structures in our components:

```
<div v-if="$slots.icon" class="p-4">
    <slot name="icon" />
</div>
```

With Storybook 7.1, these broke due to the Vue 3 renderer's updated slot generation.
Previously it used story props to populate Vue component slots, but this was
changed to use argTypes in 6c806b6, after which all stories had all slots defined.
This would cause such conditions as the example above to always be truthy.

This commit modifies the slot generation a bit to allow slots to be undefined
rather than defaulting to an `() => undefined` function value.