Do not fail when export order loader fails
The export order loader for Webpack5 currently throws if for example the babel plugin @babel/plugin-transform-modules-commonjs is in use, since it doesn't return exports in that case. The loader itself isn't crucial for Storybook to work properly. Therefore it is okay to swallow the failure and to just ignore it in cases where it fails