Added warning about lack of svelte slot support
After setting up a whole story for a component, I realized there is no way to set the slots (https://github.com/storybookjs/storybook/discussions/22774). 
All this tutorial is not useful if someone has a slot or have a plan to add a slot to their components as it will require a full re-write of the story in .svelte format. This can be avoided if we warn the user about the lack of this feature, at least until this feature is available.