Fix #26347: table-of-contents doesn't scroll and highlight
In docs page, if you click on a table-of-contents item, and there is not enough height available in the docs page, it does not change the current state of the table-of-contents.
To fix this bug a blank space is added at the end of the page to give it enough height.