Fix theming of elements inside bars
In Storybook 8.0, we have broken theming for elements inside bars.
The theming variables "barTextColor", "barHoverColor" and "barSelectedColor" were mostly
ignored and mainly other variable colors were applied.

This commit mainly ensures, that the default dark and light theming doesn't change.
If "barTextColor", "barHoverColor" or "barSelectedColor" are set by the user,
they get normally applied to the icon and text elements inside the bar areas,
like the topbar, bottom-bar or the tab area in the addons panel.