feat: migrate from express to connect
Migrates away from express, instead using connect and the underlying
node server directly.

Both express and connect support the same middleware structure, which
means most middleware packages work for both and have remained the same
in this change.

A few notes:

- Express' router has been replaced by basic connect mount points
(though the two are not equivalent exactly)
- Express static has been replaced by sirv

This trims a large sub-tree of dependencies from our packages which
express was pulling in. We in fact use very little of the functionality
express gives us, only really making use of the connect-style API
anyway.