Don’t use global Ember in @storybook/ember
The Ember global has been deprecated for quite a while now and the ember storybook framework addon relied on a hack on ember-cli-storybook to set up the global again.

This change removes the need for the Ember global by changing the way handlebars templates are precompiled and relying on a story component from ember-cli-storybook (or the user’s application) to render the actual story.

The storybook/story component can be as simple as this:

```
import Component from '@ember/component';

export default Component.extend({});
```